several updates: tauSFTool, VBS_SSWW*txts readded

+++ python/postprocessing/PrepareToPlot_tmp.py
@@ -63,11 +63,11 @@ def AreAllCondored(samplename):
     lenstore = len(storelist)
 
     if 'Data' in samplename:
-        remainder = lenstore%split
-        lenstore = lenstore/split
+        remainder = int(lenstore%split)
+        lenstore = int(lenstore/split)
         if remainder > 0:
             lenstore += 1
-
+        
     if len(condoredlist) < lenstore:
         print("condored: ", len(condoredlist), "\tlenstore: ", lenstore)
         return False
